working TweetCollector
TweetCollector is working and saves every feature, the job of accessing fields is now up to TweetNetwork. The names of the fields are now different so Tweet network class needs to be updated
